added utils balance assertions
